todo/requirements.txt
mock==1.3.0
django-smarttest==0.2.1
morelia==0.5.0
factory-boy==2.5.2
splinter[django]==0.7.3
django-coverage-plugin>=1.0b2
